Thymeleaf: cadastrar e listar

+++ src/main/java/br/com/fiap/domain/Calorias.java
@@ -2,5 +2,5 @@ package br.com.fiap.domain;
 
 public interface Calorias {
 
-	float calcularCalorias();
+	double calcularCalorias();
 }
